Extract an article callback and make a controller
