+++ server.js
@@ -1,29 +1,16 @@
-const express = require("express");
-const morgan = require("morgan");
-const { titleController } = require('./controllers');
-const DB = require("./db");
-
-const db = new DB();
+const express = require('express');
+const morgan = require('morgan');
+const { titleController, articlesController } = require('./controllers');
 
 const app = express();
 const port = 3000;
 
-app.use(morgan("tiny"));
+app.use(morgan('tiny'));
 
-app.get("/articles", (req, res) => {
-    const defaultArticle = {
-        title: 'no title',
-        body: 'no body',
-        color: 'default color',
-    }
-    const articles = db.getArticles().map(article => {
-        return {...defaultArticle, ...article};
-    })
-    res.send(articles)
-});
-app.get("/articles/title", titleController);
+app.get('/articles', articlesController);
+app.get('/articles/title', titleController);
 
 // TODO(chai): implement this controller
-app.get("/articles/body", (req, res) => res.send(''));
+app.get('/articles/body', (req, res) => res.send(''));
 
 app.listen(port, () => console.log(`Example app listening on port ${port}!`));
